Organizando os Endpoints em arquivos separados.
